create fixtures module
